Do not offer to remove redundant assignment if it is not parented by an explicit block with curly braces
Fixes #32856

For example, consider a redundant assignment to 'x' in `if (...) x = 1;`
Replacing 'x = 1' with an empty statement ';' is not useful, and user is most likely
going to remove the entire if statement in this case. However, we don't
want to suggest removing the entire if statement as that might lead to change of semantics.
So, we conservatively bail out from removable assignment analysis for such cases.