Compute relative folder information if we aren't given it
We added a new API IProjectSiteEx.AddFileEx that takes the link metadata
for an item so we don't have to query an IVsHierarchy to figure out
where the file is linked into. If we don't have a link though, we should
compute the relative folder to the project directory and use that.

Fixes https://github.com/dotnet/roslyn/issues/33171