Ensure we write logs xUnit logs and dumps in a place where we will archive them
We have two folders we use during a test run: we put our XML result
files in one folder for the Azure DevOps task to specifically catch,
and all other log files in a different folder which we upload as a
generic artifact. We were putting log files in the wrong folder so we
never uploaded them as artifacts.