Delete unnecessary code around Workspace eventing
We had some code in the VS layer that was trying to work around the
fact that we might be MEF created on a background thread, when were
didn't know what the UI thread was. It tried to assert that the first
actual workspace mutation would happen on the UI thread, and so it was
safe to create the queue at that point. This is silly now for two
reasons:

1. We're using the IThreadingContext which knows what the main thread
   is directly and without any ceremony
2. We now do indeed do workspace mutations on other threads. It's still
   the case that the first mutation is on the UI thread today but that's
   a limitation we are actively wanting to excise.