Increase diagnostics in determinism job
The determinism job has started to intermittently timeout over the last
few days. The job typically runs in ~50 minutes which is well below the
90 minute limit. On failed runs though there are no obvious outliers
that are causing the timeouts. The builds and restores are happening at
the time intervals we would expect.

This change adds more time diagontics to the run. It's unlikely that any
of these sections are responsible for the 40 minute gap we're seeing in
the bad runs but at this point need to rule out everything.