Fix deadlock with the GlobalOptionService and roaming option persister
This deadlock is a classic deadlock where two components both have a
lock, and called out to the other component while holding the lock. The
two offending methods were:

1. GlobalOptionService.GetOption, which acquires the global option lock
   and then calls into serializers. This would result in
   RoamingVisualStudioProfileOptionPersister.RecordObservedValueToWatchForChanges
   being called which would acquire the persister's lock.
2. RoamingVisualStudioProfileOptionPersister.OnSettingChangedAsync
   would acquire the persister's lock, and then call back into the
   GlobalOptionService to update it.

If both of these happened at the same time, there's a small chance of
deadlock. We've seen this sporatically in automation.

The classic fix for this is to fix one (or both) of the cases above to
avoid calling out to another component while holding the lock. Both are
easy to fix but I'm choosing (for now) to just the second case. My
reasoning is the first lock is still potentially useful: if multiple
components are requesting the same option at the same time, we will
only be calling into the settings service once rather than calling into
it multiple times simultaneously.

Fixes https://github.com/dotnet/roslyn/issues/20019.