Add telemetry to EventMap.Registry.Invoke to diagnose a crash
We're seeing crashes where it seems a workspace is opening a document
twice with no intervening close. One theory we have is something is
throwing OperationCanceledExceptions as we're notifying events,
and as a result we're not notifying everything. We've audited the
possible DocumentClose event handlers that would seem to be blamed
but we're unable to come up with any proof there is any problems there.