Fix asserts in open document handling
This code was asserting the previously-correct assertion that if a file
wasn't a regular document, it was an additional document. This updates
the asserts accordingly. We're tracking doing a full open-document
wireup in https://devdiv.visualstudio.com/DevDiv/_workitems/edit/750120,
but this is fine for now since you can still save the .editorconfig file
for it to take effect, and saving was already a requirement with our
existing implementation.