Consume updated intrinsics
The `GetPathsOfAllFilesAbove` intrinsic item function in MSBuild has
been replaced with a set of general-purpose intrinsics:
`GetPathsOfAllDirectoriesAbove`, `Combine`, and `Exists`. Here we use
these to find all the _potential_ .editorconfig files and then filter
that set to the _actual_ ones. This will allow the project systems to
pick up the set of potential files straight from MSBuild instead of
needing to recompute it themselves, and should greatly simplify their
handling of .editorconfig files.