Port Dispose analyzers (DisposeObjectsBeforeLosingScope) and (DisposableFieldsShouldBeDisposed)
We will now flag disposable creations in methods that do not escape the
method body (return value, out/ref params, assign to field/property, add
to collection, etc.). Additionally, we also flag disposable fields of
disposable types that are not disposed in the type's Dispose method.
Ported analyzers share the core dataflow analysis computation with the
analyzers in the roslyn-analyzers repo
(Microsoft.CodeAnalysis.FlowAnalysis.Utilities). This PR does not add
any code fixes. We already have a code
refactoring that offers wrapping a disposable creation within a using
statement. I am hoping to extend that to handle more cases and also
possibly convert it to a code fix in follow-up PR(s).