Workaround a CLR deadlock caused by unloading app domains with STA RCWs
If you have an STA thread that created Runtime Callable Wrappers, it's
possible to end up in a deadlock where the finalizer and app domain
unload code are waiting for each other. The suggested workaround from
the CLR team is to wait for finalizers before letting the domain
shutdown but while the thread is still running.

Fixes https://github.com/dotnet/roslyn/issues/34248