Move .editorconfig properties and items
Move .editorconfig properties and items to a separate .targets file and
import it conditionally. Automatic discovery of .editorconfig files
depends on MSBuild features that are only available in version 16.1 and
later. To avoid evaluation errors we only import the .targets file when
we're running under an MSBuild with proper support.