Reduce allocations in DiagnosticIncrementalAnalyzer
This PR reduces allocations in DiagnosticIncrementalAnalyzer by a good
margin, 450MB for loading a solution and a short edit session.

The total allocations in VSMac were 3GB, and these amounted for 450MB.

Since the methods were private, implementations can be switched from
IEnumerable and LINQ to concrete enumerator iteration and removal
of lambdas with captures.

Fixes #34761