Ensure we refresh ruleset severities after a ruleset change
If a ruleset file changed, we didn't always reapply the new values
in the IDE. For legacy projects the IDE is taking the ruleset files
and applying them to the compilation; for CPS projects that's being done
when we call into the compiler to parse the command line string. We
forgot to reparse the command line string, so CPS projects wouldn't
refresh until some other command line string changed or the project
was unloaded or reloaded.

Fixes https://github.com/dotnet/roslyn/issues/35164