Fix having the wrong syntax trees after adding an .editorconfig
Adding an .editorconfig requires us to update the compilation with
new syntax trees that contain the per-tree diagnostic options for
the compiler. When I created the CompilationTranslationAction, I
accidentally passed the old ProjectState instead of the new ProjectState
which meant we'd remove and re-add all the old trees instead of the
new trees.