Correctly handle two projects specifying the same output path
When we're converting metadata references to output references, if
there are two projects specifying the same output path, we don't do
conversion and instead just keep the reference to the metadata
reference. This was largely becasue we don't have a way to pick between
the two options, and going with "the first" isn't deterministic. There
was a bug though: when we get a second project with the same output path,
we didn't correctly do the conversion of existing references to the
first back, so you'd end up with that project reference dangling.
do the conversion back to