improve build + live dedup on full solution analysis on. (#35620)
we found a de-dup bug with full solution analysis.

basically, when we get errors, we categorize them as either syntax, semantic or other errors. it works fine for live but errors from build, we don't know what kind of errors they are. so we mark those with most common error kind (semantic) and mark that it is from build so that we can refresh them next time we do live analysis.

with full solution analysis off, it works fine since we fix those up when a file is opened. but for closed file with FSA, we sometimes fail to de-dup correctly due to our refresh optimization.

now we always refresh errors when build error is involved.