Ensure we give .editorconfig keys in the same order as in the file
Right now our naming styles .editorconfig format implicitly relies on
the order of keys being preserved from the .editorconfig file on disk
into the IDictionary<string, object> that's handed to the parser. This
is problematic in many ways, but until we either redesign the format
to avoid the need of ordering, or make the ordering explicit, we should
ensure we actually maintain that ordering.

Fixes https://github.com/dotnet/roslyn/issues/35488